Use `UpperMap` for `rgblink -S` scramble spec matching
This also makes invalid RGBLINK CLI options into fatal errors
like the other programs

+++ include/link/main.hpp
@@ -22,8 +22,8 @@ struct Options {
 	bool hasPadValue = false;
 	// Setting these three to 0 disables the functionality
 	uint16_t scrambleROMX; // -S
-	uint8_t scrambleWRAMX;
-	uint8_t scrambleSRAM;
+	uint16_t scrambleWRAMX;
+	uint16_t scrambleSRAM;
 	bool is32kMode;      // -t
 	bool beVerbose;      // -v
 	bool isWRAM0Mode;    // -w
